+++ vite.config.ts
@@ -1,3 +1,4 @@
+
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react-swc";
 import path from "path";
@@ -8,6 +9,12 @@ export default defineConfig(({ mode }) => ({
   server: {
     host: "::",
     port: 8080,
+    fs: {
+      strict: true, // Forces allowed directories ONLY
+    },
+  },
+  optimizeDeps: {
+    exclude: ['nanoid'], // Force using latest nanoid build tree
   },
   plugins: [
     react(),